# Recovery terms: use ram_file on start, but not on shutdown